Push duration and time saved handling down
